Fix renderer pointer casts

+++ neo/renderer/tr_frontend_main.cpp
@@ -76,7 +76,7 @@ void R_ToggleSmpFrame() {
 	frameData = &smpFrameData[smpFrame % NUM_FRAME_DATA];
 
 	// reset the memory allocation
-	const unsigned int bytesNeededForAlignment = FRAME_ALLOC_ALIGNMENT - ( (unsigned int)frameData->frameMemory & ( FRAME_ALLOC_ALIGNMENT - 1 ) );
+	const unsigned int bytesNeededForAlignment = FRAME_ALLOC_ALIGNMENT - ( (uintptr_t)frameData->frameMemory & ( FRAME_ALLOC_ALIGNMENT - 1 ) );
 	frameData->frameMemoryAllocated.SetValue( bytesNeededForAlignment );
 	frameData->frameMemoryUsed.SetValue( 0 );
